move stuff and format
