+++ runbot/__manifest__.py
@@ -2,22 +2,36 @@
 {
     'name': "runbot",
     'summary': "Runbot",
-    'description': "Runbot for Odoo 11.0",
+    'description': "Runbot for Odoo 13.0",
     'author': "Odoo SA",
     'website': "http://runbot.odoo.com",
     'category': 'Website',
     'version': '5.0',
     'depends': ['website', 'base', 'base_automation'],
     'data': [
-        'security/runbot_security.xml',
-        'security/ir.model.access.csv',
-        'security/ir.rule.csv',
         'data/runbot_data.xml',
         'data/runbot_build_config_data.xml',
         'data/build_parse.xml',
         'data/runbot_error_regex_data.xml',
         'data/error_link.xml',
         'data/website_data.xml',
+
+        'security/runbot_security.xml',
+        'security/ir.model.access.csv',
+        'security/ir.rule.csv',
+
+        'templates/assets.xml',
+        'templates/badge.xml',
+        'templates/batch.xml',
+        'templates/build.xml',
+        'templates/bundle.xml',
+        'templates/commit.xml',
+        'templates/dashboard.xml',
+        'templates/frontend.xml',
+        'templates/git.xml',
+        'templates/nginx.xml',
+        'templates/utils.xml',
+
         'views/assets.xml',
         'views/repo_views.xml',
         'views/branch_views.xml',
@@ -31,17 +45,8 @@
         'views/stat_views.xml',
         'views/warning_views.xml',
         'views/commit_views.xml',
+
         'wizards/mutli_build_wizard_views.xml',
         'wizards/stat_regex_wizard_views.xml',
-        'templates/frontend.xml',
-        'templates/build.xml',
-        'templates/assets.xml',
-        'templates/dashboard.xml',
-        'templates/nginx.xml',
-        'templates/git.xml',
-        'templates/badge.xml',
-        'templates/bundle.xml',
-        'templates/commit.xml',
-
     ],
 }
