+++ runbot/templates/assets.xml
@@ -1,10 +1,10 @@
 <odoo>
-    <data>
-      <template id="assets_frontend" inherit_id="website.assets_frontend" name="runbot.assets.frontend">
-        <xpath expr="." position="inside">
-          <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/octicons/2.0.2/octicons.css"/>
-          <script type="text/javascript" src="/runbot/static/src/js/runbot.js"/>
-        </xpath>
-      </template>
-    </data>
-  </odoo>
+  <data>
+    <template id="assets_frontend" inherit_id="website.assets_frontend" name="runbot.assets.frontend">
+      <xpath expr="." position="inside">
+        <link rel="stylesheet" href="/runbot/static/src/css/runbot.css"/>
+        <script type="text/javascript" src="/runbot/static/src/js/runbot.js"/>
+      </xpath>
+    </template>
+  </data>
+</odoo>
