--- runbot/views/assets.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <template id="assets_front_end" inherit_id="web.assets_frontend" name="runbot assets">
-            <xpath expr="." position="inside">
-                <link rel="stylesheet" href="/runbot/static/src/css/runbot.css"/>
-            </xpath>
-        </template>
-    </data>
-</odoo>
